refactor(ivy): move core code to core.ts (#23476)
PR Close #23476
